spider/requirements.txt
anyio==3.6.1
astroid==2.11.6
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==37.0.4
decorator==5.1.1
dill==0.3.5.1
future==0.18.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
nest-asyncio==1.5.5
numpy==1.23.0
outcome==1.2.0
platformdirs==2.5.2
pycparser==2.21
pylint==2.14.3
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tomli==2.0.1
tomlkit==0.11.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.9
validators==0.18.2
weaviate-client==3.5.1
wrapt==1.14.1
wsproto==1.1.0
